+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.4.0
+
+### 🍀 Added
+
+- Added a new **Blocktad** drawing format. It works similarly to **Octads**, but uses blocks instead of braille characters
+
+### 🛠 Changed
+
+- Improved stability and performance of the diffing step (by @airblast-dev)
+- Removed unnecessary allocations (by @airblast-dev)
+
+### ⚠ Breaking
+
+- Removed `Color::NO_COLOR` in favor of `Attributes::NO_FG_COLOR` and `Attributes::NO_BG_COLOR`
+
 # 0.3.0
 
 ### 🍀 Added